feat(serializer): 实现了多种序列化方式和自定义序列化

+++ rpc-common/src/main/java/com/rickey/rpc/common/service/UserService.java
@@ -18,9 +18,10 @@ public interface UserService {
 
     /**
      * test mock
+     *
      * @return number
      */
-    default short getNumber(){
+    default int getNumber() {
         return 1;
     }
 }
